% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badgr-package.R
\docType{package}
\name{badgr-package}
\alias{badgr}
\alias{badgr-package}
\title{badgr: Create A shields.io Badge URL}
\description{
Build a URL for a custom shields.io badge to use in your project's README file or elsewhere.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
