% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{get_badge_static}
\alias{get_badge_static}
\title{Generate The URL Endpoint For A Static Metadata Badge From shields.io}
\usage{
get_badge_static(
  label = "shields.io",
  message = "badge",
  color = "ff0000",
  logo_path = NULL,
  logo_filetype = NULL,
  include_md = TRUE,
  to_clipboard = TRUE,
  open_browser = TRUE
)
}
\arguments{
\item{label}{Text to display on the left half of the badge.}

\item{message}{Text to display on the right half of the badge.}

\item{color}{The color of the left side of the badge.}

\item{logo_path}{The URL of the logo image.}

\item{logo_filetype}{The filetype of the logo image.}

\item{include_md}{Do you want to prepare the URL with Markdown syntax to allow
for direct copying into a Markdown file? Otherwise you get the bare URL.}

\item{to_clipboard}{Do you want the string to be copied to your clipboard so
you can paste it elsewhere? Will overwrite your current clipboard items.}

\item{open_browser}{Do you want to preview the badge in the browser? Requires
an internet connection.}
}
\value{
A text string.
}
\description{
Generate The URL Endpoint For A Static Metadata Badge From shields.io
}
\examples{
\dontrun{
get_badge_static("Label", "Message", "blue")
}
}
