% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badgr-package.R
\docType{package}
\name{badgr-package}
\alias{badgr}
\alias{badgr-package}
\title{badgr: Create A shields.io Badge URL}
\description{
Build A URL For A Custom shields.io Badge To Use In You Project's README File Or Elsewhere.
}
\author{
\strong{Maintainer}: Matt Dray \email{mwdray@gmail.com}

}
\keyword{internal}
