% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{get_badge}
\alias{get_badge}
\title{Generate A Static shields.io Badge}
\usage{
get_badge(
  label = "shields.io",
  message = "badge",
  color = "ff0000",
  style = "flat",
  label_color = NULL,
  md_link = NULL,
  link_left = NULL,
  link_right = NULL,
  logo_simple = NULL,
  logo_color = NULL,
  logo_width = NULL,
  logo_path = NULL,
  ext_override = NULL,
  browser_preview = TRUE,
  include_md = TRUE,
  to_clipboard = TRUE
)
}
\arguments{
\item{label}{Text string to display on the left side of the badge.}

\item{message}{Text string to display on the right side of the badge.}

\item{color}{The background color of the right side of the badge. Should be
one of hex, RGB, RGBA, HSL, HSLA or a CSS named color.}

\item{style}{One of \code{"plastic"}, \code{"flat"}, \code{"for-the-badge"}
or \code{"social"}. See examples on \href{https://shields.io/}{shields.io}.}

\item{label_color}{The background color of the left side of the badge. Should
be one of hex, RGB, RGBA, HSL, HSLA or a CSS named color.}

\item{md_link}{A URL string to be used in your badge's Markdown output. If
used, clicking the rendered HTML badge output will take you to this site.}

\item{link_left}{A URL string for the left side of the badge. (Included for
completeness; you probably want \code{md_link} instead)}

\item{link_right}{A URL string for the right side of the badge. (Included for
completeness; you probably want \code{md_link} instead)}

\item{logo_simple}{Text string that names a logo from
\href{https://simpleicons.org/}{simple-icons.org} to be placed on the
far left of the badge. (Use either a named logo with \code{logo_simple}
or a custom logo with \code{logo_path}; not both.)}

\item{logo_color}{A color for the logo specified by \code{logo_simple}.}

\item{logo_width}{Horizontal spacing width around the logo.}

\item{logo_path}{URL string for your custom logo image. (Use either a named
logo with \code{logo_simple} or a custom logo with \code{logo_path}; not
both.)}

\item{ext_override}{Provide a file extension type for your logo_path that
overrides the one in \code{logo_path}.}

\item{browser_preview}{Do you want to preview the badge in the browser? Requires
an internet connection.}

\item{include_md}{Do you want to prepare the URL with Markdown syntax
(\code{![]()}) to allow for direct copying into a Markdown file?
Otherwise you get the bare URL. Use \code{md_link} to add a link to the
badge.}

\item{to_clipboard}{Do you want the string to be copied to your clipboard so
you can paste it elsewhere? Will overwrite your current clipboard items.}
}
\value{
A text string.
}
\description{
Build a URL string that generates a \href{https://shields.io/}{shields.io}
metadata badge. For convenience, this function allows you to embed the URL in
Markdown with a link and copies this to the clipboard for use in a README or
elsewhere.
}
\examples{
get_badge("Label", "Message", "blue", browser_preview = FALSE, to_clipboard = FALSE)
}
